requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
chromadb==0.4.15
pypdf==3.17.0
pdfplumber==0.9.0
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.0
requests==2.31.0
plotly==5.17.0
pandas==2.1.3
sentence-transformers==2.2.2
numpy==1.24.3
gunicorn==21.2.0
